leetcode/1901-2000/1971.Find-if-Path-Exists-in-Graph/Solution_test.go
package Solution

import (
	"reflect"
	"strconv"
	"testing"
)

func TestSolution(t *testing.T) {
	//	测试用例
	cases := []struct {
		name   string
		n int
		edges [][]int
		source, destination int
		expect bool
	}{
		{"TestCase1", 3, [][]int{
		{0, 1},
		{1, 2},
		{2, 0},
	}, 0, 2, true},
		{"TestCase2", 6, [][]int{
		{0, 1},
		{0, 2},
		{3, 5},
		{5, 4},
		{4, 3},
	}, 0, 5, false},
	}

	//	开始测试
	for i, c := range cases {
		t.Run(c.name+" "+strconv.Itoa(i), func(t *testing.T) {
			got := Solution(c.n, c.edges, c.source, c.destination)
			if !reflect.DeepEqual(got, c.expect) {
				t.Fatalf("expected: %v, but got: %v, with inputs: %v %v %v %v",
					c.expect, got, c.n, c.edges, c.source, c.destination)
			}
			got = Solution1(c.n, c.edges, c.source, c.destination)
			if !reflect.DeepEqual(got, c.expect) {
				t.Fatalf("expected: %v, but got: %v, with inputs: %v %v %v %v",
					c.expect, got, c.n, c.edges, c.source, c.destination)
			}
		})
	}
}

//	压力测试
func BenchmarkSolution(b *testing.B) {
}

//	使用案列
func ExampleSolution() {
}
